Module3 [Extra exercises]/SK-Agent connecting to MCP/requirements.txt
semantic-kernel
python-dotenv
httpx
fastapi
uvicorn
mcp
